core: support saving boot transaction marker on success
Similar to the method through which suricatta allows configuring its
behavior with the bootloader, allow the boot transaction marker's
behavior to be configured as an aid to double-copy implementations.

If the boot transaction marker is configured to be saved in the bootloader
environment after a successful A/B update, the bootloader then only needs
to check for the presence of this marker on the next boot, and remove it
once its has selected the correct boot partition.

Persistence of the transaction marker allows the bootloader's partition
selection implementation to exist independent of the sw-description in a
particular .swu file.  It allows the bootloader to distinguish from a
standard power-on reset where no update was performed (the last booted
partition should be booted again after reset), versus a reset following
an update (where the boot partition should be switched).